Drop support for Python 2.7 and 3.5

+++ glue/utils/qt/colors.py
@@ -1,5 +1,3 @@
-from __future__ import absolute_import, division, print_function
-
 import numpy as np
 from matplotlib.colors import ColorConverter
 
@@ -39,7 +37,7 @@ def mpl_to_qt_color(color, alpha=None):
     r, g, b, a = cc.to_rgba(color)
     if alpha is not None:
         a = alpha
-    return QtGui.QColor(r * 255, g * 255, b * 255, a * 255)
+    return QtGui.QColor(int(r * 255), int(g * 255), int(b * 255), int(a * 255))
 
 
 def qt_to_mpl_color(qcolor):
